Refactor: Student関連・CourseMaster関連のクラスを役割ごとにパッケージ分割、受講生・コース管理APIのOpenAPI仕様（Swagger）を追加
