+++ src/main/java/raisetech/student/management/controller/student/StudentController.java
@@ -14,24 +14,23 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import raisetech.student.management.controller.StudentSearchForm;
-import raisetech.student.management.data.Course;
 import raisetech.student.management.domain.StudentDetail;
-import raisetech.student.management.service.StudentService;
+import raisetech.student.management.service.student.StudentService;
 
 /**
  * 受講生の検索や登録、更新などを行うREST APIとして受け付けるControllerです。
  */
 @Validated
 @RestController
-@Tag(name = "受講生管理", description = "受講生の登録・検索・更新を行うAPI")
+@RequestMapping("/students")
+@Tag(name = "受講生管理", description = "受講生情報の登録・検索・更新を行うAPI")
 public class StudentController {
 
   private final StudentService service;
@@ -58,7 +57,7 @@ public class StudentController {
       @ApiResponse(responseCode = "404", description = "該当する受講生が見つかりません"),
       @ApiResponse(responseCode = "500", description = "サーバーエラー")
   })
-  @GetMapping("/students")  // GET /students?__=__
+  @GetMapping  // GET /students?__=__
   public ResponseEntity<List<StudentDetail>> searchStudents(StudentSearchForm searchForm) {
 
     List<StudentDetail> studentDetails = service.getStudentList(searchForm.toDTO());
@@ -81,7 +80,7 @@ public class StudentController {
       @ApiResponse(responseCode = "400", description = "更新処理に失敗しました"),
       @ApiResponse(responseCode = "500", description = "サーバーエラー")
   })
-  @PostMapping("/students")
+  @PostMapping
   public ResponseEntity<StudentDetail> registerStudent(
       @RequestBody @Valid StudentDetail studentDetail) {
 
@@ -103,7 +102,7 @@ public class StudentController {
       @ApiResponse(responseCode = "404", description = "該当する受講生が見つかりません"),
       @ApiResponse(responseCode = "500", description = "サーバーエラー")
   })
-  @GetMapping("/students/{studentId}")
+  @GetMapping("/{studentId}")
   public StudentDetail getStudent(
       @Parameter(description = "検索する受講生のID", example = "12")
       @PathVariable @Valid Integer studentId) {
@@ -122,7 +121,7 @@ public class StudentController {
       @ApiResponse(responseCode = "400", description = "更新処理に失敗しました"),
       @ApiResponse(responseCode = "500", description = "サーバーエラー")
   })
-  @PutMapping("/students")
+  @PutMapping
   public ResponseEntity<String> updateStudent(
       @RequestBody @Validated({Default.class, Update.class}) StudentDetail studentDetail) {
     service.updateStudent(studentDetail);
@@ -130,98 +129,6 @@ public class StudentController {
         studentDetail.getStudent().getFullName() + " さんの更新処理が成功しました。");
     // 更新が完了したらレスポンスとしてOK(200)とメッセージを返す。
   }
-
-  /**
-   * 【コースマスタの全件取得】開講状況にかかわらず全て表示
-   *
-   * @return コースマスタリスト（全件）
-   */
-  @Operation(summary = "コースマスタ取得", description = "コースマスタを全件取得します")
-  @ApiResponses(value = {
-      @ApiResponse(responseCode = "200", description = "成功"),
-      @ApiResponse(responseCode = "404", description = "コースマスタが見つかりません"),
-      @ApiResponse(responseCode = "500", description = "サーバーエラー")
-  })
-  @GetMapping("/courses")
-  public ResponseEntity<List<Course>> getCourses() {
-    var courses = service.getCourseList();
-    if (courses.isEmpty()) {
-      return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-    }
-    return ResponseEntity.ok(courses);
-  }
-
-  /**
-   * 【コースマスタの登録】POSTで受け取った情報を元に新規マスタ登録を行う。
-   *
-   * @param course 入力情報（コースマスタ情報）
-   * @return 登録されたコースマスタ情報（自動生成されたIDを含む）
-   */
-  @Operation(summary = "コースマスタ登録", description = "新規コースマスタを登録します。")
-  @ApiResponses(value = {
-      @ApiResponse(responseCode = "201", description = "成功"),
-      @ApiResponse(responseCode = "400", description = "登録処理に失敗しました"),
-      @ApiResponse(responseCode = "500", description = "サーバーエラー")
-  })
-  @PostMapping("/courses")
-  public ResponseEntity<Course> registerCourseMaster(@RequestBody @Valid Course course) {
-
-    service.registerCourseMaster(course);
-    // 登録されたIDから、登録されたマスタ情報を呼び出し
-    var createdCourse = service.getCourseMaster(course.getCourseId());
-    URI location = URI.create("/courses/" + createdCourse.getCourseId());
-
-    return ResponseEntity.created(location).body(createdCourse);
-  }
-
-  /**
-   * 【コースマスタの更新】受け取った情報をもとに既存のコースマスタを更新する。
-   *
-   * @param course 入力情報（コースマスタ情報）
-   */
-  @Operation(summary = "コースマスタ更新", description = "既存のコースマスタを更新します。")
-  @ApiResponses(value = {
-      @ApiResponse(responseCode = "200", description = "成功"),
-      @ApiResponse(responseCode = "400", description = "更新処理に失敗しました"),
-      @ApiResponse(responseCode = "404", description = "更新対象コースマスタが見つかりません"),
-      @ApiResponse(responseCode = "500", description = "サーバーエラー"),
-  })
-  @PutMapping("/courses")
-  public ResponseEntity<String> updateCourseMaster(
-      @RequestBody @Validated({Default.class, Update.class}) Course course) {
-    service.updateCourseMaster(course);
-    return ResponseEntity.ok(
-        "コース名：【 " + course.getCourseName() + " 】の更新処理が成功しました。");
-  }
-
-  /**
-   * 【コースマスタの削除】
-   *
-   * @param courseId 削除対象のコースマスタID
-   */
-  @Operation(summary = "コースマスタ削除", description = "既存のコースマスタを削除します。")
-  @ApiResponses(value = {
-      @ApiResponse(responseCode = "204", description = "コースマスタ削除に成功しました"),
-      @ApiResponse(responseCode = "400", description = "削除に失敗しました"),
-      @ApiResponse(responseCode = "404", description = "削除対象コースマスタが見つかりません"),
-      @ApiResponse(responseCode = "500", description = "サーバーエラー"),
-  })
-  @DeleteMapping("/courses/{courseId}")
-  public ResponseEntity<String> deleteCourseMaster(@PathVariable @Valid Integer courseId) {
-    service.deleteCourseMaster(courseId);
-    return ResponseEntity.noContent().build();
-  }
-
-  /**
-   * 【リダイレクト】 旧URIが指定された場合に、正しいURI"/students"にリダイレクトする。
-   *
-   * @return /students リダイレクト
-   */
-  @Operation(summary = "/students リダイレクト", description = "旧URIのハンドリングメソッドです。")
-  @GetMapping("/studentList")
-  public ResponseEntity<Void> redirectStudents() {
-    return ResponseEntity.status(HttpStatus.SEE_OTHER)
-        .location(URI.create("/students")).build();
-  }
-
+  
 }
+
