+++ src/main/java/raisetech/student/management/repository/student/StudentRepository.java
@@ -1,9 +1,7 @@
 package raisetech.student.management.repository.student;
 
-import jakarta.validation.Valid;
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
-import raisetech.student.management.data.Course;
 import raisetech.student.management.data.CourseStatus;
 import raisetech.student.management.data.Student;
 import raisetech.student.management.data.StudentsCourse;
@@ -13,7 +11,7 @@ import raisetech.student.management.dto.StudentSearchDTO;
  * 受講生情報を扱うRepository（インターフェース） 受講生テーブルと受講生コース情報テーブルと紐づいています。 全件検索や単一条件での検索、コース情報の検索が行えるクラスです。
  */
 @Mapper
-public interface CourseRepository {
+public interface StudentRepository {
 
   /**
    * 【受講生の一覧表示】
@@ -125,40 +123,5 @@ public interface CourseRepository {
    * @return 該当するステータス情報リスト
    */
   List<CourseStatus> findStatus(StudentSearchDTO condition);
-
-  /**
-   * 【コースマスタの全件取得】
-   *
-   * @return コースマスタリスト
-   */
-  List<Course> displayCourseMaster();
-
-  /**
-   * 【コースマスタの新規登録】
-   *
-   * @param course 登録内容（コースマスタ情報）
-   */
-  int registerCourseMaster(@Valid Course course);
-
-  /**
-   * 【コースマスタのID検索】
-   *
-   * @param courseId 検索するID
-   * @return 該当するコースマスタ
-   */
-  Course searchCourseMaster(Integer courseId);
-
-  /**
-   * 【コースマスタの更新】
-   *
-   * @param course 更新内容（コースマスタ）
-   */
-  int updateCourseMaster(Course course);
-
-  /**
-   * 【コースマスタの削除】
-   *
-   * @param courseId 削除対象のコースID
-   */
-  void deleteCourseMaster(@Valid Integer courseId);
+  
 }
