+++ src/main/java/raisetech/student/management/service/student/StudentService.java
@@ -215,73 +215,5 @@ public class StudentService {
     }
 
   }
-
-//  /**
-//   * 【コースマスタの全件取得】
-//   *
-//   * @return コースマスタリスト（全件）
-//   */
-//  public List<Course> getCourseList() {
-//    return studentRepository.displayCourseMaster();
-//  }
-//
-//  /**
-//   * 【コースマスタの新規作成】
-//   *
-//   * @param course 入力情報（マスタ情報）
-//   */
-//  public void registerCourseMaster(Course course) {
-//    studentRepository.registerCourseMaster(course);
-//  }
-//
-//  /**
-//   * 【コースマスタのID検索】
-//   *
-//   * @param courseId 検索ID
-//   * @return 該当するコースマスタ
-//   */
-//  public Course getCourseMaster(Integer courseId) {
-//    return studentRepository.searchCourseMaster(courseId);
-//  }
-//
-//  /**
-//   * 【コースマスタの更新】
-//   *
-//   * @param course 入力された更新情報（コースマスタ）
-//   */
-//  public void updateCourseMaster(Course course) {
-//
-//    // 事前に対象のコースマスタを検索（なければ例外throw）
-//    Integer targetCourseId = course.getCourseId();
-//    Course courseExist = studentRepository.searchCourseMaster(targetCourseId);
-//    if (courseExist == null) {
-//      throw new NoDataException(
-//          "更新対象のコースマスタが見つかりません。[ID: " + targetCourseId + " ]"
-//      );
-//    }
-//
-//    int updateMasterData = studentRepository.updateCourseMaster(course);
-//
-//    // 更新されなかった場合に例外をスルー
-//    if (updateMasterData == 0) {
-//      throw new ProcessFailedException("コースマスタは更新されませんでした。");
-//    }
-//
-//  }
-//
-//  /**
-//   * 【コースマスタの削除】
-//   *
-//   * @param courseId 削除対象のコースID
-//   */
-//  public void deleteCourseMaster(@Valid Integer courseId) {
-//
-//    // 事前に対象のコースマスタを検索（なければ例外throw）
-//    Course courseExist = studentRepository.searchCourseMaster(courseId);
-//    if (courseExist == null) {
-//      throw new NoDataException("削除対象が見つかりません。[ID: " + courseId + " ]");
-//    }
-//
-//    studentRepository.deleteCourseMaster(courseId);
-//  }
+  
 }
