+++ src/test/java/raisetech/student/management/controller/student/StudentControllerTest.java
@@ -3,16 +3,13 @@ package raisetech.student.management.controller.student;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import java.sql.Timestamp;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyInt;
 import org.mockito.Mockito;
-import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -27,9 +24,6 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import raisetech.student.management.data.Course;
-import raisetech.student.management.data.Course.CourseCategory;
-import raisetech.student.management.data.Course.CourseName;
 import raisetech.student.management.data.CourseStatus;
 import raisetech.student.management.data.CourseStatus.Status;
 import raisetech.student.management.data.Student;
@@ -37,7 +31,7 @@ import raisetech.student.management.data.StudentsCourse;
 import raisetech.student.management.domain.CourseDetail;
 import raisetech.student.management.domain.StudentDetail;
 import raisetech.student.management.exception.NoDataException;
-import raisetech.student.management.service.course.StudentService;
+import raisetech.student.management.service.student.StudentService;
 
 @WebMvcTest(StudentController.class)
 @Import(StudentControllerTest.MockConfig.class) // モックBeanを定義したクラスをインポート
@@ -53,7 +47,6 @@ class StudentControllerTest {
   private CourseDetail courseDetail1;
   private CourseDetail courseDetail2;
   private StudentDetail studentDetail;
-  private Course course;
 
   // 日時固定
   LocalDateTime fixedDateTime = LocalDateTime.of(2021, 5, 7, 16, 0, 0);
@@ -87,9 +80,9 @@ class StudentControllerTest {
 
     studentDetail = new StudentDetail(student, List.of(courseDetail1, courseDetail2));
 
-    course = new Course(
-        1, CourseName.Javaコース, CourseCategory.開発系コース, 6, false,
-        Timestamp.valueOf("2021-05-07 16:00:00"), Timestamp.valueOf("2021-05-07 16:00:00"));
+//    course = new Course(
+//        1, CourseName.Javaコース, CourseCategory.開発系コース, 6, false,
+//        Timestamp.valueOf("2021-05-07 16:00:00"), Timestamp.valueOf("2021-05-07 16:00:00"));
 
     Mockito.reset(service); // モックをリセット
 
@@ -256,96 +249,6 @@ class StudentControllerTest {
         .andExpect(status().isNotFound())
         .andExpect(
             content().string("受講生情報が見つかりませんでした。ID: 1234567890"));  // エラーメッセージの確認
-  }
-
-  @Test
-  void コースマスタ全件取得_正常完了_200OKとマスタリストが返ってくること()
-      throws Exception {
-    when(service.getCourseList()).thenReturn(List.of(course));
-
-    // リクエストを想定
-    mockMvc.perform(MockMvcRequestBuilders.get("/courses")
-            .contentType(MediaType.APPLICATION_JSON))
-        .andExpect(status().isOk())  // ステータスコード200
-        .andExpect(
-            content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-        .andExpect(jsonPath("$[0].courseId").value(1))
-        .andExpect(jsonPath("$[0].courseName").value("Javaコース"))
-        .andExpect(jsonPath("$[0].category").value("開発系コース"))
-        .andExpect(jsonPath("$[0].createdAt").value("2021-05-07T07:00:00.000+00:00"))
-        .andExpect(jsonPath("$[0].updatedAt").value("2021-05-07T07:00:00.000+00:00"))
-    ;
-
-    // serviceメソッドが呼ばれたことを確認
-    verify(service, times(1)).getCourseList();
-  }
-
-  @Test
-  void コースマスタ全件取得_コースマスタが存在しない場合_404NotFoundが返ること()
-      throws Exception {
-    when(service.getCourseList()).thenReturn(List.of());
-
-    mockMvc.perform(MockMvcRequestBuilders.get("/courses")
-            .contentType(MediaType.APPLICATION_JSON))
-        .andExpect(status().isNotFound());
-
-    verify(service, times(1)).getCourseList();
-  }
-
-  @Test
-  void コースマスタ登録_正常完了_201Createdと登録された受講生情報が返ってくること()
-      throws Exception {
-
-    var objectMapper = new ObjectMapper()
-        .registerModule(new JavaTimeModule())
-        .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-    String request = objectMapper.writeValueAsString(course);
-
-    when(service.getCourseMaster(anyInt())).thenReturn(course);
-
-    mockMvc.perform(MockMvcRequestBuilders.post("/courses")
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(request))
-        .andExpect(status().isCreated())
-        .andExpect(jsonPath("$.courseId").value("1"))
-        .andExpect(jsonPath("$.courseName").value("Javaコース"))
-        .andExpect(jsonPath("$.duration").value("6"))
-        .andExpect(jsonPath("$.closed").value("false"))
-        .andExpect(jsonPath("$.createdAt").value("2021-05-07T07:00:00.000+00:00"))
-        .andExpect(jsonPath("$.updatedAt").value("2021-05-07T07:00:00.000+00:00"))
-    ;
-
-    verify(service, times(1)).registerCourseMaster(any());
-    verify(service, times(1)).getCourseMaster(anyInt());
-  }
-
-  @Test
-  void コースマスタ更新_正常完了_200OKと実行結果が返ってくること() throws Exception {
-
-    var objectMapper = new ObjectMapper()
-        .registerModule(new JavaTimeModule())
-        .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-    String request = objectMapper.writeValueAsString(course);
-
-    mockMvc.perform(MockMvcRequestBuilders.put("/courses")
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(request))
-        .andExpect(status().isOk())
-        .andExpect(content().string("コース名：【 Javaコース 】の更新処理が成功しました。"));
-
-    verify(service, times(1)).updateCourseMaster(any());
-  }
-
-  @Test
-  void コースマスタ削除_正常完了_204NoContentを返すこと() throws Exception {
-
-    Integer courseId = 1;
-    doNothing().when(service).deleteCourseMaster(courseId);
-
-    mockMvc.perform(MockMvcRequestBuilders.delete("/courses/{courseId}", courseId))
-        .andExpect(status().isNoContent());
-
-    verify(service, times(1)).deleteCourseMaster(courseId);
   }
 
 }
